+++ packages/ubuntu_desktop_installer/lib/app.dart
@@ -8,7 +8,7 @@ import 'app_theme.dart';
 import 'l10n/app_localizations.dart';
 import 'pages/allocate_disk_space_page.dart';
 import 'pages/choose_your_look_page.dart';
-import 'pages/keyboard_layout_page.dart';
+import 'pages/keyboard_layout/keyboard_layout_page.dart';
 import 'pages/try_or_install/try_or_install_page.dart';
 import 'pages/turn_off_rst_page.dart';
 import 'pages/updates_other_software/updates_other_software_page.dart';
